// .PRD/erp-refactor/v3/types/stock-entry-detail.ts
import { z } from 'zod';

export const StockEntryDetailSchema = z.object({
  id: z.string(),
  owner: z.string().optional(),
  creation: z.string().optional(),
  modified: z.string().optional(),
  modified_by: z.string().optional(),
  barcode: z.string().optional(),
  has_item_scanned: z.boolean().optional().default(false),
  s_warehouse: z.string().optional(),
  t_warehouse: z.string().optional(),
  item_code: z.string(),
  item_name: z.string().optional(),
  is_finished_item: z.boolean().optional().default(false),
  is_scrap_item: z.boolean().optional().default(false),
  quality_inspection: z.string().optional(),
  subcontracted_item: z.string().optional(),
  against_fg: z.string().optional(),
  description: z.string().optional(),
  item_group: z.string().optional(),
  image: z.string().optional(),
  image_view: z.string().optional(),
  qty: z.number(),
  transfer_qty: z.number(),
  retain_sample: z.boolean().optional().default(false),
  uom: z.string(),
  stock_uom: z.string(),
  conversion_factor: z.number(),
  sample_quantity: z.number().int().optional(),
  basic_rate: z.number().optional(),
  customer_provided_item_cost: z.number().optional(),
  additional_cost: z.number().optional(),
  landed_cost_voucher_amount: z.number().optional(),
  valuation_rate: z.number().optional(),
  allow_zero_valuation_rate: z.boolean().optional().default(false),
  set_basic_rate_manually: z.boolean().optional().default(false),
  basic_amount: z.number().optional(),
  amount: z.number().optional(),
  use_serial_batch_fields: z.boolean().optional().default(false),
  serial_and_batch_bundle: z.string().optional(),
  serial_no: z.string().optional(),
  batch_no: z.string().optional(),
  expense_account: z.string().optional(),
  cost_center: z.string().optional().default(':Company'),
  project: z.string().optional(),
  actual_qty: z.number().optional(),
  transferred_qty: z.number().optional(),
  bom_no: z.string().optional(),
  allow_alternative_item: z.boolean().optional().default(false),
  material_request: z.string().optional(),
  material_request_item: z.string().optional(),
  original_item: z.string().optional(),
  against_stock_entry: z.string().optional(),
  ste_detail: z.string().optional(),
  po_detail: z.string().optional(),
  sco_rm_detail: z.string().optional(),
  scio_detail: z.string().optional(),
  putaway_rule: z.string().optional(),
  reference_purchase_receipt: z.string().optional(),
  job_card_item: z.string().optional(),
});

export type StockEntryDetail = z.infer<typeof StockEntryDetailSchema>;

export const StockEntryDetailInsertSchema = StockEntryDetailSchema.omit({ id: true, owner: true, creation: true, modified: true, modified_by: true });
export type StockEntryDetailInsert = z.infer<typeof StockEntryDetailInsertSchema>;